# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EDOLE AFRICA ADMIN - Unified operational management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Auth tokens and user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedClients"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client with contacts and summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /clients/{id}/contacts:
    get:
      operationId: listClientContacts
      tags: [clients]
      summary: List contacts for a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientContact"
    post:
      operationId: createClientContact
      tags: [clients]
      summary: Create a contact for a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientContactBody"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientContact"

  # ── CRM ────────────────────────────────────────────────────────────────────
  /crm/opportunities:
    get:
      operationId: listOpportunities
      tags: [crm]
      summary: List CRM opportunities
      parameters:
        - in: query
          name: stage
          schema:
            type: string
        - in: query
          name: clientId
          schema:
            type: string
        - in: query
          name: assignedTo
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated opportunities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOpportunities"
    post:
      operationId: createOpportunity
      tags: [crm]
      summary: Create a new opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpportunityBody"
      responses:
        "201":
          description: Created opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
  /crm/opportunities/{id}:
    get:
      operationId: getOpportunity
      tags: [crm]
      summary: Get opportunity details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Opportunity details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
    put:
      operationId: updateOpportunity
      tags: [crm]
      summary: Update an opportunity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpportunityBody"
      responses:
        "200":
          description: Updated opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
    delete:
      operationId: deleteOpportunity
      tags: [crm]
      summary: Delete an opportunity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /crm/activities:
    get:
      operationId: listActivities
      tags: [crm]
      summary: List CRM activities
      parameters:
        - in: query
          name: clientId
          schema:
            type: string
        - in: query
          name: opportunityId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated activities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedActivities"
    post:
      operationId: createActivity
      tags: [crm]
      summary: Create a CRM activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityBody"
      responses:
        "201":
          description: Created activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"

  # ── PROJECTS ───────────────────────────────────────────────────────────────
...
  /collaborators:
    get:
      operationId: listCollaborators
      tags: [collaborators]
      summary: List all collaborators
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: available
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated collaborators
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCollaborators"
    post:
      operationId: createCollaborator
      tags: [collaborators]
      summary: Create a collaborator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollaboratorBody"
      responses:
        "201":
          description: Created collaborator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collaborator"
  /collaborators/{id}:
    get:
      operationId: getCollaborator
      tags: [collaborators]
      summary: Get collaborator details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collaborator details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collaborator"
    put:
      operationId: updateCollaborator
      tags: [collaborators]
      summary: Update a collaborator
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollaboratorBody"
      responses:
        "200":
          description: Updated collaborator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collaborator"
    delete:
      operationId: deleteCollaborator
      tags: [collaborators]
      summary: Delete a collaborator
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        avatarUrl:
          type: string
        isClient:
          type: boolean
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isClient

    # ── USER ─────────────────────────────────────────────────────────────────
...
    Client:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        industry:
          type: string
        address:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        status:
          type: string
          enum: [active, inactive, prospect]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - createdAt
    ClientDetail:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            contacts:
              type: array
              items:
                $ref: "#/components/schemas/ClientContact"
            projectsCount:
              type: integer
            ordersCount:
              type: integer
            totalRevenue:
              type: number
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        industry:
          type: string
        address:
          type: string
        website:
          type: string
        status:
          type: string
          enum: [active, inactive, prospect]
      required:
        - name
    ClientContact:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        isPrimary:
          type: boolean
      required:
        - id
        - clientId
        - firstName
        - lastName
    CreateClientContactBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        isPrimary:
          type: boolean
      required:
        - firstName
        - lastName
    PaginatedClients:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ── CRM ──────────────────────────────────────────────────────────────────
    Opportunity:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        stage:
          type: string
          enum: [lead, qualified, proposal, negotiation, won, lost]
        value:
          type: number
        currency:
          type: string
        probability:
          type: integer
        assignedToId:
          type: string
        assignedToName:
          type: string
        expectedCloseDate:
          type: string
          format: date
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - stage
        - createdAt
    CreateOpportunityBody:
      type: object
      properties:
        title:
          type: string
        clientId:
          type: string
        stage:
          type: string
          enum: [lead, qualified, proposal, negotiation, won, lost]
        value:
          type: number
        currency:
          type: string
        probability:
          type: integer
        assignedToId:
          type: string
        expectedCloseDate:
          type: string
          format: date
        notes:
          type: string
      required:
        - title
        - stage
...
    Activity:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [call, email, meeting, note, task]
        subject:
          type: string
        description:
          type: string
        clientId:
          type: string
        opportunityId:
          type: string
        userId:
          type: string
        userName:
          type: string
        scheduledAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - subject
        - createdAt
    CreateActivityBody:
      type: object
      properties:
        type:
          type: string
          enum: [call, email, meeting, note, task]
        subject:
          type: string
        description:
          type: string
        clientId:
          type: string
        opportunityId:
          type: string
        scheduledAt:
          type: string
          format: date-time
      required:
        - type
        - subject
    PaginatedOpportunities:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Opportunity"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    PaginatedActivities:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ── PROJECT ───────────────────────────────────────────────────────────────
...
    Collaborator:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        position:
          type: string
        department:
          type: string
        avatarUrl:
          type: string
        isAvailable:
          type: boolean
        currentProjectsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - createdAt
    CreateCollaboratorBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        position:
          type: string
        department:
          type: string
        isAvailable:
          type: boolean
      required:
        - firstName
        - lastName
...
    PaginatedCollaborators:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Collaborator"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ── EQUIPMENT ─────────────────────────────────────────────────────────────
...